refactor: enum 파일 경로 이동(#80)
