+++ Kkuk/Presentation/SearchContent/View/SearchContentViewController.swift
@@ -7,25 +7,6 @@
 
 import UIKit
 
-enum SegmentMenu: Int {
-    case title
-    case memo
-
-    var name: String {
-        switch self {
-        case .title: return "사이트"
-        case .memo: return "메모"
-        }
-    }
-
-    var column: String {
-        switch self {
-        case .title: return "title"
-        case .memo: return "memo"
-        }
-    }
-}
-
 class SearchContentViewController: BaseUIViewController {
     
     var contentList: [Content] = []
